Create separate tree_of_boxes module, reshuffle docs
