+++ boxtree/tree_build.py
@@ -4,25 +4,17 @@
 Building Particle-Based Trees
 -----------------------------
 
+These functions produce instances of the particle-based :class:`Tree`.
+
+.. note::
+
+    These functions currently keep their bulk data in in
+    :class:`pyopencl.array.Array` instances.  This contrasts with the box-based
+    tree (:class:`TreeOfBoxes`), which operates on data  in :class:`numpy.ndarray`
+    instances. Along with the rest of :mod:`boxtree`, both will migrate to
+    :mod:`arraycontext` in the future.
+
 .. autoclass:: TreeBuilder
-
-.. _tree-of-boxes:
-
-Manipulating Trees of Boxes
----------------------------
-
-These functions manipulate instances of :class:`TreeOfBoxes`.
-
-These functions currently keep their bulk data in :class:`numpy.ndarray` instances,
-however, along with the rest of :mod:`boxtree`, this will migrate to
-:mod:`arraycontext` in the future.
-
-.. autofunction:: make_tob_root
-.. autofunction:: refined
-.. autofunction:: uniformly_refined
-.. autofunction:: coarsened
-.. autofunction:: refined_and_coarsened
-.. autofunction:: make_mesh_from_leaves
 """
 
 __copyright__ = """
@@ -51,20 +43,13 @@ THE SOFTWARE.
 """
 
 
-from typing import Optional, Tuple, TYPE_CHECKING
-import sys
-
 import numpy as np
-import numpy.typing as npt
 from pytools import memoize_method
 import pyopencl as cl
 import pyopencl.array  # noqa
 from functools import partial
-from boxtree.tree import Tree, TreeOfBoxes
+from boxtree.tree import Tree
 from pytools import ProcessLogger, DebugProcessLogger
-
-if TYPE_CHECKING or sys._BUILDING_SPHINX_DOCS:
-    from meshmode.mesh import Mesh
 
 import logging
 logger = logging.getLogger(__name__)
@@ -74,281 +59,7 @@ class MaxLevelsExceeded(RuntimeError):
     pass
 
 
-# {{{ utils for tree of boxes
-
-def _resized_array(arr: npt.NDArray, new_size: int) -> npt.NDArray:
-    """Return a resized copy of the array. The new_size is a scalar which is
-    applied to the last dimension.
-    """
-    old_size = arr.shape[-1]
-    prefix = (slice(None), ) * (arr.ndim - 1)
-    if old_size >= new_size:
-        return arr[prefix + (slice(new_size), )].copy()
-    else:
-        new_shape = list(arr.shape)
-        new_shape[-1] = new_size
-        new_arr = np.zeros(new_shape, arr.dtype)
-        new_arr[prefix + (slice(old_size), )] = arr
-        return new_arr
-
-
-def _vec_of_signs(dim: int, i: int) -> npt.NDArray:
-    """The sign vector is obtained by converting i to a dim-bit binary.
-    """
-    # e.g. bin(10) = '0b1010'
-    binary_digits = [int(bd) for bd in bin(i)[2:]]
-    n = len(binary_digits)
-    assert n <= dim
-    return np.array([0]*(dim-n) + binary_digits) * 2 - 1
-
-
-# {{{ refine/coarsen a tree of boxes
-
-def refined(tob: TreeOfBoxes, refine_flags: npt.NDArray) -> TreeOfBoxes:
-    """Make a refined copy of `tob` where boxes flagged with `refine_flags` are
-    refined.
-    """
-    return refined_and_coarsened(tob, refine_flags, None)
-
-
-def uniformly_refined(tob: TreeOfBoxes) -> TreeOfBoxes:
-    """Make a uniformly refined copy of `tob`.
-    """
-    refine_flags = np.zeros(tob.nboxes, bool)
-    refine_flags[tob.leaf_flags] = 1
-    return refined(tob, refine_flags)
-
-
-def coarsened(tob: TreeOfBoxes, coarsen_flags: npt.NDArray,
-              error_on_ignored_flags: bool = True
-              ) -> TreeOfBoxes:
-    """Make a coarsened copy of `tob` where boxes flagged with `coarsen_flags`
-    are coarsened.
-    """
-    return refined_and_coarsened(
-        tob, None, coarsen_flags, error_on_ignored_flags)
-
-
-def _apply_refine_flags_without_sorting(refine_flags, tob):
-    if refine_flags[~tob.leaf_flags].any():
-        raise ValueError("attempting to split non-leaf")
-
-    refine_parents, = np.where(refine_flags)
-    if len(refine_parents) == 0:
-        return tob
-
-    nchildren = 2**tob.dim
-    n_new_boxes = len(refine_parents) * nchildren
-    nboxes_new = tob.nboxes + n_new_boxes
-
-    child_box_starts = (
-            tob.nboxes
-            + nchildren * np.arange(len(refine_parents)))
-
-    refine_parents_per_child = np.empty(
-            (nchildren, len(refine_parents)), np.intp)
-    refine_parents_per_child[:] = refine_parents.reshape(-1)
-    refine_parents_per_child = refine_parents_per_child.reshape(-1)
-
-    box_parents = _resized_array(tob.box_parent_ids, nboxes_new)
-    box_centers = _resized_array(tob.box_centers, nboxes_new)
-    box_children = _resized_array(tob.box_child_ids, nboxes_new)
-    box_levels = _resized_array(tob.box_levels, nboxes_new)
-
-    # new boxes are appended at the end, so applying coarsen_flags wrt the
-    # original tree is still meaningful after this
-    box_parents[tob.nboxes:] = refine_parents_per_child
-    box_levels[tob.nboxes:] = tob.box_levels[box_parents[tob.nboxes:]] + 1
-    box_children[:, refine_parents] = (
-        child_box_starts + np.arange(nchildren).reshape(-1, 1))
-
-    for i in range(2**tob.dim):
-        children_i = box_children[i, refine_parents]
-        offsets = (
-                tob.root_extent * _vec_of_signs(tob.dim, i).reshape(-1, 1)
-                * (1/2**(1+box_levels[children_i])))
-        box_centers[:, children_i] = (
-                box_centers[:, refine_parents] + offsets)
-
-    return TreeOfBoxes(
-        box_centers=box_centers, root_extent=tob.root_extent,
-        box_parent_ids=box_parents, box_child_ids=box_children,
-        box_levels=box_levels)
-
-
-def _apply_coarsen_flags(coarsen_flags, tob, error_on_ignored_flags=True):
-    if coarsen_flags[~tob.leaf_flags].any():
-        raise ValueError("attempting to coarsen non-leaf")
-    coarsen_sources, = np.where(coarsen_flags)
-    if not coarsen_sources:
-        return tob
-
-    coarsen_parents = tob.box_parent_ids[coarsen_sources]
-    coarsen_peers = tob.box_child_ids[:, coarsen_parents].reshape(-1)
-    coarsen_peer_is_leaf = tob.leaf_flags[coarsen_peers]
-    coarsen_exec_flags = np.all(coarsen_peer_is_leaf, axis=0)
-
-    # when a leaf box marked for coarsening has non-leaf peers
-    coarsen_flags_ignored = (coarsen_exec_flags != coarsen_flags)
-    if np.any(coarsen_flags_ignored):
-        msg = (f"{np.sum(coarsen_flags_ignored)} out of "
-               f"{np.sum(coarsen_flags)} coarsening flags ignored "
-               "to prevent removing non-leaf boxes")
-        if error_on_ignored_flags:
-            raise RuntimeError(msg)
-        else:
-            import warnings
-            warnings.warn(msg)
-
-    # deleted boxes are marked as:
-    # level = inf
-    # parent = -1
-    coarsen_parents = coarsen_parents[coarsen_exec_flags]
-    coarsen_peers = coarsen_peers[:, coarsen_exec_flags]
-    box_parents = tob.box_parent_ids.copy()
-    box_parents[coarsen_peers] = -1
-    box_children = tob.box_child_ids.copy()
-    box_children[:, coarsen_parents] = 0
-    box_levels = tob.box_levels.copy()
-    box_levels[coarsen_peers] = np.inf
-
-    from dataclasses import replace
-    return replace(tob, box_parent_ids=box_parents,
-                   box_child_ids=box_children, box_levels=box_levels)
-
-
-def _sort_boxes_by_level(tob, queue=None):
-    if np.any(np.diff(tob.box_levels) < 0):
-        # reorder boxes to into non-decreasing levels
-        neworder = np.argsort(tob.box_levels)
-        box_centers = tob.box_centers[:, neworder]
-        box_parent_ids = tob.box_parent_ids[neworder]
-        box_child_ids = tob.box_child_ids[:, neworder]
-        box_levels = tob.box_levels[neworder]
-        tob = TreeOfBoxes(
-            box_centers=box_centers, root_extent=tob.root_extent,
-            box_parent_ids=box_parent_ids, box_child_ids=box_child_ids,
-            box_levels=box_levels)
-    return tob
-
-
-def _sort_and_prune_deleted_boxes(tob):
-    tob = _sort_boxes_by_level(tob)
-    n_stale_boxes = np.sum(tob.box_levels == np.inf)
-    newn = tob.nboxes - n_stale_boxes
-    from dataclasses import replace
-    return replace(tob, box_parent_ids=tob.box_parent_ids[:newn],
-                   box_child_ids=tob.box_child_ids[:, :newn],
-                   box_levels=tob.box_levels[:newn],
-                   box_centers=tob.box_centers[:, :newn])
-
-
-def refined_and_coarsened(tob: TreeOfBoxes,
-                          refine_flags: Optional[npt.NDArray] = None,
-                          coarsen_flags: Optional[npt.NDArray] = None,
-                          error_on_ignored_flags: bool = True
-                          ) -> TreeOfBoxes:
-    """Make a refined/coarsened copy. When children of the same parent box
-    are marked differently, the refinement flag takes priority.
-
-    Both refinement and coarsening flags can only be set of leaves.
-    To prevent drastic mesh change, coarsening is only executed when a leaf
-    box is marked for coarsening, and its parent's children are all leaf
-    boxes (so that change in the number of boxes is bounded per box flagged).
-    Please note that the above behavior may be subject to change in the future.
-
-    :arg refine_flags: a boolean array of size `nboxes`.
-    :arg coarsen_flags: a boolean array of size `nboxes`.
-    :arg error_on_ignored_flags: if true, an exception is raised when enforcing
-        level restriction requires ignoring some coarsening flags.
-    :returns: a processed copy of the tree.
-    """
-    if refine_flags is None:
-        refine_flags = np.zeros(tob.nboxes, bool)
-    if coarsen_flags is None:
-        coarsen_flags = np.zeros(tob.nboxes, bool)
-
-    if (refine_flags & coarsen_flags).any():
-        raise ValueError("some boxes are simultaneously marked "
-                         "to refine and coarsen")
-
-    tob = _apply_refine_flags_without_sorting(refine_flags, tob)
-    coarsen_flags = _resized_array(coarsen_flags, tob.nboxes)
-    tob = _apply_coarsen_flags(coarsen_flags, tob, error_on_ignored_flags)
-    return _sort_and_prune_deleted_boxes(tob)
-
-# }}} End refine/coarsen a tree of boxes
-
-
-def make_tob_root(bbox: npt.NDArray) -> TreeOfBoxes:
-    """
-    Make the minimal tree of boxes, consisting of a single root box filling
-    *bbox*.
-
-    :arg bbox: a 2-by-dim numpy array of ``[lower_bounds, upper_bounds]`` for the
-        bounding box.
-
-    .. note::
-
-        *bbox* is expected to be square (with tolerances as accepted by
-        :func:`numpy.allclose`).
-    """
-    from pytools import single_valued
-    assert len(bbox) == 2
-    dim = single_valued([len(bbox[0]), len(bbox[1])])
-    center = np.array(
-        [(bbox[0][iaxis] + bbox[1][iaxis]) * 0.5 for iaxis in range(dim)])
-    extents = np.array(
-        [(bbox[1][iaxis] - bbox[0][iaxis]) for iaxis in range(dim)])
-    from pytools import single_valued
-    parents = np.array([0], np.intp)
-    parents[0] = -1  # root has no parent
-    return TreeOfBoxes(
-            box_centers=center.reshape(dim, 1),
-            root_extent=single_valued(extents, np.allclose),
-            box_parent_ids=parents,
-            box_child_ids=np.array([0] * 2**dim, np.intp).reshape(2**dim, 1),
-            box_levels=np.array([0]),
-            )
-
-
-def make_mesh_from_leaves(tob: TreeOfBoxes) -> Tuple["Mesh", np.ndarray]:
-    """Make a :class:`~meshmode.mesh.Mesh` from the leaf boxes of the tree
-    of boxes *tob*.
-
-    :returns: A tuple of the mesh and a vector of the element number -> box number
-        mapping.
-    """
-    lfboxes = tob.leaf_boxes
-    lfcenters = tob.box_centers[:, lfboxes]
-    lflevels = tob.box_levels[lfboxes]
-    lfradii = tob.root_extent / 2 / (2**lflevels)
-
-    # use tensor product nodes ordering
-    import modepy.nodes as nd
-    cell_nodes_1d = np.array([-1, 1])
-    cell_nodes = nd.tensor_product_nodes(tob.dim, cell_nodes_1d)
-
-    lfvertices = (
-        np.repeat(lfcenters, 2**tob.dim, axis=1)
-        + np.repeat(lfradii, 2**tob.dim) * np.tile(cell_nodes, (1, len(lfboxes)))
-    )
-
-    # FIXME: purge redundant vertices
-    from meshmode.mesh import Mesh, TensorProductElementGroup
-    from meshmode.mesh.generation import make_group_from_vertices
-
-    vertex_indices = np.arange(
-        len(lfboxes) * 2**tob.dim, dtype=np.int32).reshape([-1, 2**tob.dim])
-    group = make_group_from_vertices(
-        lfvertices, vertex_indices, 1,
-        group_cls=TensorProductElementGroup,
-        unit_nodes=None)
-
-    return Mesh(vertices=lfvertices, groups=[group]), tob.leaf_boxes
-
-# }}}
-
+# {{{ tree builder
 
 class TreeBuilder:
     """
@@ -2087,4 +1798,6 @@ class TreeBuilder:
 
     # }}}
 
-# vim: foldmethod=marker:filetype=pyopencl
+# }}}
+
+# vim: foldmethod=marker
