+++ doc/tree-build.rst
@@ -1,0 +1,7 @@
+Building Trees
+==============
+
+.. automodule:: boxtree.tree_of_boxes
+
+.. automodule:: boxtree.tree_build
+
