+++ doc/tree.rst
@@ -1,7 +1,6 @@
-Tree building
-=============
+Tree Data Structures
+====================
 
 .. automodule:: boxtree.tree
-.. automodule:: boxtree.tree_build
 
 .. vim: sw=4
